+++ lib/aws-sdk-rails.rb
@@ -12,9 +12,9 @@ require_relative 'aws/rails/sqs_active_job/lambda_handler'
 require_relative 'aws/rails/middleware/ebs_sqs_active_job_middleware'
 
 require_relative 'action_dispatch/session/dynamodb_store'
-require_relative 'active_job/queue_adapters/amazon_sqs_adapter'
-require_relative 'active_job/queue_adapters/amazon_sqs_adapter/params'
-require_relative 'active_job/queue_adapters/amazon_sqs_async_adapter'
+require_relative 'active_job/queue_adapters/sqs_adapter'
+require_relative 'active_job/queue_adapters/sqs_adapter/params'
+require_relative 'active_job/queue_adapters/sqs_async_adapter'
 
 require_relative 'generators/aws_record/base'