+++ spec/active_job/queue_adapters/sqs_adapter/params_spec.rb
@@ -4,7 +4,7 @@ require 'test_helper'
 
 module ActiveJob
   module QueueAdapters
-    class AmazonSqsAdapter
+    class SqsAdapter
       describe Params do
         describe '.assured_delay_seconds' do
           let(:now) { Time.now }