+++ spec/dummy/config/application.rb
@@ -12,6 +12,6 @@ module Dummy
     config.eager_load = false
     config.require_master_key = true
 
-    config.active_job.queue_adapter = :amazon
+    config.active_job.queue_adapter = :sqs
   end
 end