+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Unreleased Changes
 
 * Issue - Ensure `:ses` or `:sesv2` as ActionMailer configuration.
 
+* Feature - Do not allow `:amazon`, `amazon_sqs`, or `amazon_sqs_async` for SQS active job configuration. Instead use `:sqs` and `:sqs_async`.
+
+
 3.13.0 (2024-06-06)
 ------------------